Create extensions in the configuration
This is thought to be a very basic test, whether the configuration of
extensions is working. If something is very wrong, OSIAM would not start
and all tests would fail. Unfortunately, we cannot test, whether the
extensions have really been created, because DBUnit clears the complete
database on every test run.